ignition-cmake: clean up package_info() a bit

+++ recipes/ignition-cmake/all/conanfile.py
@@ -51,23 +51,21 @@ class IgnitionCmakeConan(ConanFile):
         cmake.install()
         rmdir(self, os.path.join(self.package_folder, "lib", "pkgconfig"))
         rmdir(self, os.path.join(self.package_folder, "bin"))
+
         version_major = Version(self.version).major
         cmake_config_files_dir = os.path.join(self.package_folder, "lib", "cmake", f"ignition-cmake{version_major}")
-        files = os.listdir(cmake_config_files_dir)
-
-        # retain the special config file for utilities target provided by ignition-cmake
-        # removing it from the list
-        files.remove(f"ignition-cmake{version_major}-utilities-targets.cmake")
-
-        # remove all other xxx.cmake files from the list
-        for file in files:
+        for file in os.listdir(cmake_config_files_dir):
             if file.endswith(".cmake"):
+                if file == f"ignition-cmake{version_major}-utilities-targets.cmake":
+                    # retain the special config file for utilities target provided by ignition-cmake
+                    continue
                 os.remove(os.path.join(cmake_config_files_dir, file))
 
-        # add version information for downstream dependencies consuming ign-cmake through cmake_find_package generators
-        self._create_cmake_module_variables(os.path.join(self.package_folder, self._module_file_rel_path), Version(self.version))
-
-
+        # add version information for downstream dependencies consuming ign-cmake through CMake generators
+        self._create_cmake_module_variables(
+            os.path.join(self.package_folder, self._module_file_rel_path),
+            Version(self.version)
+        )
 
     def _create_cmake_module_variables(self, module_file, version):
         # the version info is needed by downstream ignition-dependencies
@@ -87,38 +85,37 @@ class IgnitionCmakeConan(ConanFile):
     def package_info(self):
         self.cpp_info.bindirs = []
         self.cpp_info.libdirs = []
+        self.cpp_info.frameworkdirs = []
 
         version_major = Version(self.version).major
         ign_cmake_component = f"ignition-cmake{version_major}"
-        base_module_path = os.path.join(self.package_folder, "lib", "cmake", ign_cmake_component)
+        self.cpp_info.set_property("cmake_file_name", ign_cmake_component)
+
+        base_module_path = os.path.join("lib", "cmake", ign_cmake_component)
         ign_cmake_file = os.path.join(base_module_path, f"cmake{version_major}", "IgnCMake.cmake")
         utils_targets_file = os.path.join(base_module_path, f"{ign_cmake_component}-utilities-targets.cmake")
+        self.cpp_info.set_property("cmake_build_modules", [self._module_file_rel_path, ign_cmake_file, utils_targets_file])
 
-        self.cpp_info.set_property("cmake_file_name", ign_cmake_component)
-        self.cpp_info.set_property("cmake_target_name", f"{ign_cmake_component}::{ign_cmake_component}")
+        self.cpp_info.components[ign_cmake_component].set_property("cmake_target_name", f"{ign_cmake_component}::{ign_cmake_component}")
+        self.cpp_info.components[ign_cmake_component].builddirs.append(os.path.join(base_module_path, f"cmake{version_major}"))
+
+        self.cpp_info.components["utilities"].set_property("cmake_target_name", f"{ign_cmake_component}::utilities")
+        self.cpp_info.components["utilities"].builddirs.append(os.path.join(base_module_path, f"cmake{version_major}"))
+        self.cpp_info.components["utilities"].includedirs.append(os.path.join("include", "ignition", f"cmake{version_major}"))
 
         # TODO: to remove in conan v2 once cmake_find_package_* generators removed
         self.cpp_info.names["cmake_find_package"] = ign_cmake_component
         self.cpp_info.names["cmake_find_package_multi"] = ign_cmake_component
         self.cpp_info.names["cmake_paths"] = ign_cmake_component
-
-        self.cpp_info.components[ign_cmake_component].set_property("cmake_target_name", f"{ign_cmake_component}::{ign_cmake_component}")
         self.cpp_info.components[ign_cmake_component].names["cmake_find_package"] = ign_cmake_component
         self.cpp_info.components[ign_cmake_component].names["cmake_find_package_multi"] = ign_cmake_component
         self.cpp_info.components[ign_cmake_component].names["cmake_paths"] = ign_cmake_component
-        self.cpp_info.components[ign_cmake_component].builddirs.append(os.path.join(base_module_path, f"cmake{version_major}"))
-
         self.cpp_info.components[ign_cmake_component].build_modules["cmake_find_package"] = [self._module_file_rel_path, ign_cmake_file]
         self.cpp_info.components[ign_cmake_component].build_modules["cmake_find_package_multi"] = [self._module_file_rel_path, ign_cmake_file]
         self.cpp_info.components[ign_cmake_component].build_modules["cmake_paths"] = [self._module_file_rel_path, ign_cmake_file]
-
-        self.cpp_info.components["utilities"].set_property("cmake_target_name", f"{ign_cmake_component}::utilities")
         self.cpp_info.components["utilities"].names["cmake_find_package"] = "utilities"
         self.cpp_info.components["utilities"].names["cmake_find_package_multi"] = "utilities"
         self.cpp_info.components["utilities"].names["cmake_paths"] = "utilities"
-        self.cpp_info.components["utilities"].includedirs.append(f"include/ignition/cmake{version_major}")
-
-        self.cpp_info.set_property("cmake_build_modules", [self._module_file_rel_path, ign_cmake_file])
         self.cpp_info.components["utilities"].build_modules["cmake_find_package"] = [self._module_file_rel_path, ign_cmake_file, utils_targets_file]
         self.cpp_info.components["utilities"].build_modules["cmake_find_package_multi"] = [self._module_file_rel_path, ign_cmake_file, utils_targets_file]
         self.cpp_info.components["utilities"].build_modules["cmake_paths"] = [self._module_file_rel_path, ign_cmake_file, utils_targets_file]
